dept(typescript) : migrate new command to typescript
