+++ lib/package-managers/index.ts
@@ -1,2 +1,3 @@
 export * from './package-manager';
 export * from './package-manager.factory';
+export * from './abstract.package-manager';
