+++ package.json
@@ -34,6 +34,7 @@
     "os-name": "^2.0.1"
   },
   "devDependencies": {
+    "@types/inquirer": "0.0.41",
     "@types/node": "^9.6.6",
     "@types/node-emoji": "^1.8.0",
     "@types/ora": "^1.3.4",
